backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
google-api-python-client
groq
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
aiosqlite
supabase
